Add generic scale (up or down) function
Previously we could only scale up by number of workers.
To scale down we needed to specify the precise workers to kill
(this was normally handled by the scheduler in adaptive mode).
However sometimes you just want to manually control the size of the
cluster.

This implements a simple heuristic, to first choose the idle workers
then then prefer workers with less memory use.