Automatically read worker yaml file from env variable
This allows users to launch a KubeCluster() with no arguments if the
worker pod spec is in a yaml file pointed to by
the DASKERNETES_WORKER_SPEC_PATH environment variable.

This commit also creates a config.py file that contains daskernetes
configuration, which is currently just taken from environment variables.