Set the name of each worker to allow tracking worker removals.
Before this change, `KubeCluster._update_worker_status` was not working correctly. The name found in `KubeCluster.scheduler_info` defaulted to worker address, and was thus different from the corresponding key in `KubeCluster.workers`. After this change, the auto-generated worker name is handled correctly by the `Worker` and is appended to `args` in the Pod spec.